new endpoints

+++ src/main/java/com/example/teststrategy/repositories/LoginRepository.java
@@ -3,5 +3,5 @@ package com.example.teststrategy.repositories;
 import com.example.teststrategy.models.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<Login, Long> {
+public interface LoginRepository extends JpaRepository<Login, Integer> {
 }
